recipe_service/requirements.txt
asyncpg==0.30.0
sqlalchemy==2.0.43
pydantic_settings==2.10.1
asyncio==4.0.0

#Testing dependencies
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.0
